@@ -1,6 +1,6 @@
 {
   "name": "jammusic",
-  "version": "1.4.5",
+  "version": "1.5.0",
   "description": "joshandmariamusic.com",
   "main": "dist/index.html",
   "repository": {
@@ -9,7 +9,8 @@
   },
   "author": "https://github.com/orgs/WebJamApps/teams/handlersandkickers",
   "engines": {
-    "node": ">=16.8.0"
+    "node": ">=16.13.0 <17",
+    "npm": ">=8.1.0 <9"
   },
   "files": [
     "webpack.config.js",
@@ -32,7 +33,7 @@
     "test:lint-fix": "npm run test:stylelint && nps test.lint.fix",
     "test:lint": "npm run test:stylelint && nps test.lint",
     "test": "npm run test:lint && npm run test:jest",
-    "test:local": "npm run test:lint-fix && npm run test:jest && npm run test:puppeteer && npm run cc",
+    "test:local": "npm run test:lint-fix && npm run test:jest && npm run cc",
     "test:jest": "jest --config jest.config.json --no-cache --runInBand",
     "test:puppeteer": "jest --preset jest-puppeteer --config jest.puppeteer-config.json /test/puppeteer --runInBand --forceExit",
     "cleaninstall": "rm -rf yarn.lock && rm -rf package-lock.json && rm -rf node_modules && npm cache clean --force && npm install",
@@ -40,8 +41,10 @@
     "test:stylelint": "stylelint \"static/**/*.scss\""
   },
   "dependencies": {
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
     "@material-ui/core": "latest",
-    "@material-ui/icons": "latest",
+    "@mui/material": "^5.1.0",
     "@popperjs/core": "^2.9.2",
     "@tinymce/tinymce-react": "^3.12.6",
     "@types/copy-webpack-plugin": "latest",
@@ -88,7 +91,7 @@
     "mini-css-extract-plugin": "^1.3.1",
     "moment": "latest",
     "mui-datatables": "latest",
-    "nps": "latest",
+    "nps": "^5.9.8",
     "nps-utils": "^1.0.2",
     "popper.js": "latest",
     "process": "^0.11.10",
