@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import React from 'react';
 import { shallow } from 'enzyme';
-import { Button, Select } from '@material-ui/core';
+import { Button } from '@mui/material';
 import Inquiry from '../../../src/components/Inquiry';
 
 describe('Inquiry Form', () => {
@@ -14,12 +14,6 @@ describe('Inquiry Form', () => {
   });
   it('renders correctly', () => {
     expect(wrapper).toMatchSnapshot();
-  });
-  it('displays the usa state dropdown when country selected is United States', () => {
-    wrapper.instance().setState({ country: 'United States' });
-    const dropdown = wrapper.find(Select).get(1);
-    dropdown.props.onChange({ target: { value: 'Virginia' } });
-    expect(wrapper.instance().state.uSAstate).toBe('Virginia');
   });
   it('calls on change for text input', () => {
     wrapper.instance().setState = jest.fn();
