@@ -3,16 +3,16 @@ import MUIDataTable from 'mui-datatables';
 import type { Dispatch, AnyAction } from 'redux';
 import { ISong, SongsContext } from '../../providers/Songs.provider';
 import songsTableUtils from './songsTableUtils';
+import { EditorContext } from '../../providers/Editor.provider';
 
 type Props = {
   sData?:ISong[], token:string, dispatch: Dispatch<AnyAction>
 };
-const SongsTable:React.FC<Props> = ({ sData, token, dispatch }): JSX.Element => {
-  const { test, songs } = React.useContext(SongsContext);
+const SongsTable:React.FC<Props> = ({ sData, token }): JSX.Element => {
+  const { songs } = React.useContext(SongsContext);
+  const { editor, setNewEditor } = React.useContext(EditorContext);
   let tableData = sData !== undefined ? sData : songs;
-  tableData = tableData.length > 0 ? songsTableUtils.addButtons(tableData, token, dispatch) : [];
-  // eslint-disable-next-line no-console
-  console.log(test);
+  tableData = tableData.length > 0 ? songsTableUtils.addButtons(tableData, token, setNewEditor, editor) : [];
   return (
     <div style={{ padding: '10px', marginBottom: '-20px' }}>
       {tableData.length > 0 ? (
