@@ -16,7 +16,7 @@ interface MusicDashboardProps extends RouteComponentProps<Record<string, string 
   dispatch: Dispatch<AnyAction>;
   scc: AGClientSocket;
   auth: { token: string };
-  editPic: Iimage | { _id: string, title: string, url: string, thumbnail: string };
+  editPic: Iimage;
   editSong: ISong | { _id: string, category: string, year: number, title: string, url: string },
   editTour: { date?: string; time?: string; tickets?: string; more?: string; venue?: string; location?: string; _id?: string; datetime?: string };
   showTable: boolean;
@@ -77,7 +77,6 @@ export class MusicDashboard extends Component<MusicDashboardProps, MusicDashboar
     this.editTourAPI = this.editTourAPI.bind(this);
     this.resetEditForm = this.resetEditForm.bind(this);
     this.onChangeSong = this.onChangeSong.bind(this);
-    this.handleCategoryChange = this.handleCategoryChange.bind(this);
     this.setSongState = this.setSongState.bind(this);
     this.handleNavClick = this.handleNavClick.bind(this);
     this.handleRadioChange = this.handleRadioChange.bind(this);
@@ -87,7 +86,7 @@ export class MusicDashboard extends Component<MusicDashboardProps, MusicDashboar
 
   componentDidUpdate(prevProps:MusicDashboardProps): void {
     const { editSong } = this.props;
-    if (editSong._id !== prevProps.editSong._id) { this.setSongState(editSong); }
+    if (editSong && editSong._id !== prevProps.editSong._id) { this.setSongState(editSong); }
   }
 
   onChange(evt: React.ChangeEvent<HTMLInputElement>): void {
@@ -112,12 +111,6 @@ export class MusicDashboard extends Component<MusicDashboardProps, MusicDashboar
 
   setFormTime(time: string): void { this.setState({ time }); }
 
-  // eslint-disable-next-line react/sort-comp
-  handleCategoryChange(event: React.ChangeEvent<HTMLSelectElement>): void {
-    const { songState } = this.state;
-    this.setState({ songState: { ...songState, category: event.target.value } });
-  }
-
   handleEditorChange(venue: string): void { this.setState({ venue }); this.checkEdit(); }
 
   // eslint-disable-next-line react/sort-comp
@@ -135,6 +128,8 @@ export class MusicDashboard extends Component<MusicDashboardProps, MusicDashboar
 
   handleRadioChange(evt: { target: { value: string } }): void {
     this.setState({ showCaption: evt.target.value });
+    const { dispatch } = this.props;
+    dispatch({ type: 'EDIT_PIC', data: { ...this.props.editPic, comments:evt.target.value } });
   }
 
   // eslint-disable-next-line class-methods-use-this
