@@ -13,13 +13,14 @@ export interface ISong {
   image?: string;
   title: string;
   url: string;
-  _id: string;
+  _id?: string;
   modify?:JSX.Element
 }
 
 export const SongsContext = createContext({
   test: '',
   songs: [defaultSong],
+  resetSongs:/*istanbul ignore next */(...args:any)=>{},
 });
 type Props = { children: ReactChild };
 
@@ -27,11 +28,12 @@ export const SongsProvider = ({ children }: Props): JSX.Element => {
   const { Provider } = SongsContext;
   const [test] = useState('the songs provider has been successfully connected :)');
   const [songs, setSongs] = useSongsState<ISong[]>([defaultSong]);
+  const resetSongs = setSongs;
   useEffect(() => {
     fetchSongs.getSongs(setSongs);
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);// if we don't include this as empty array, it keeps repeatedly fetching the songs
-  return (<Provider value={{ test, songs }}>{children}</Provider>
+  return (<Provider value={{ test, songs, resetSongs }}>{children}</Provider>
   );
 };
 
